roadcast/requirements.txt
flask>=2.0
torch>=1.13
torchvision>=0.14
Pillow>=9.0
tqdm>=4.60
